Refactor: move stories under story directory
